[Firewall] clean rules on exit by default
Bofre we left the rules active after NetTrust was shutdown
This was a problem for cases were we want to continue
communicating with the internet after we stop/remove
NetTrust.

Now we have enabled by default a cleanup rule. On exit
NetTrust will remove rules from chains and then the chain

If we want to keep communication filtered after NetTrust
has exited, then we can pass -do-not-flush-table on NetTrust

With -do-not-flush-table or doNotFlushTable from config file
NetTrust will keep rules on exit but still will empty the
authorized set (this was also the case, even before we allowed
to clean up the chain)